scripts/requirements.txt
openai-batch-sdk
python-dotenv
